# api/openapi.yaml
openapi: 3.1.0
info:
  title: Ydisks 闲鱼助手 API
  version: 1.0.0
  description: 版本化 HTTP API 的唯一契约源。
servers:
  - url: /
paths:
...
  "/api/v1/admin/cookies":
    get:
      operationId: getApiV1AdminCookies
      summary: getApiV1AdminCookies
      security: [{cookieAuth: []}]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: [array, 'null']
                items:
                  $ref: '#/components/schemas/AdminCookie'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  "/api/v1/analytics/dashboard":
    get:
      operationId: getApiV1AnalyticsDashboard
      summary: getApiV1AnalyticsDashboard
      security: [{cookieAuth: []}]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardStatsResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  "/api/v1/analytics/orders":
    get:
      operationId: getApiV1AnalyticsOrders
      summary: getApiV1AnalyticsOrders
      security: [{cookieAuth: []}]
      parameters:
        - name: start_date
          in: query
          required: true
          schema: {type: string, format: date}
        - name: end_date
          in: query
          required: true
          schema: {type: string, format: date}
        - name: timezone_offset_minutes
          in: query
          required: false
          schema: {type: integer}
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  "/api/v1/automation-rules":
    get:
      operationId: getApiV1AutomationRules
      summary: getApiV1AutomationRules
      security: [{cookieAuth: []}]
      parameters:
        - name: page
          in: query
          required: false
          schema: {type: integer, minimum: 1}
        - name: page_size
          in: query
          required: false
          schema: {type: integer, minimum: 1, maximum: 100}
        - name: cookie_id
          in: query
          required: false
          schema: {type: string}
        - name: trigger_type
          in: query
          required: false
          schema: {type: string}
        - name: enabled
          in: query
          required: false
          schema: {type: boolean}
        - name: search
          in: query
          required: false
          schema: {type: string}
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutomationRulesResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: postApiV1AutomationRules
      summary: postApiV1AutomationRules
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutomationRuleRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 配置冲突或发货模板状态已变化
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  "/api/v1/cards":
    get:
      operationId: getApiV1Cards
      summary: getApiV1Cards
      security: [{cookieAuth: []}]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CardResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: postApiV1Cards
      summary: postApiV1Cards
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardMutationRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MutationIDResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  "/api/v1/default-replies":
    get:
      operationId: getApiV1DefaultReplies
      summary: getApiV1DefaultReplies
      security: [{cookieAuth: []}]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultReplyMapResponse'
        '400':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 统一错误响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    AutomationRuleRequest:
      type: object
      required: [cookie_id, item_id, name, trigger_type, enabled, priority, config_json, actions]
      properties:
        cookie_id: {type: string}
        item_id: {type: string}
        name: {type: string}
        trigger_type: {type: string}
        enabled: {type: boolean}
        priority: {type: integer}
        config_json: {type: string, description: '规则 JSON 配置；账号级 order_paid 仅在 allow_all_items 为布尔 true 时允许全商品兜底，省略或 false 均不授权。'}
        actions: {type: array, items: {$ref: '#/components/schemas/AutomationActionRequest'}}
...
    AutomationRulesResponse:
      oneOf:
        - $ref: '#/components/schemas/AutomationRuleListResponse'
        - $ref: '#/components/schemas/AutomationRulePageResponse'
...
    MutationIDResponse:
      type: object
      required: [success, id]
      properties:
        success: {type: boolean}
        id: {type: integer}
...
    AdminCookie:
      type: object
      required: [id, user_id, remark, created_at, owner, enabled]
      properties:
        id: {type: string}
        user_id: {type: integer}
        remark: {type: string}
        created_at: {type: string}
        owner: {type: string}
        enabled: {type: boolean}
...
    DashboardStatsResponse:
      type: object
      required: [total_cookies, active_cookies, total_cards, available_card_stock, total_keywords, total_orders]
      properties:
        total_cookies: {type: integer}
        active_cookies: {type: integer}
        total_cards: {type: integer}
        available_card_stock: {type: integer}
        total_keywords: {type: integer}
        total_orders: {type: integer}
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        details:
          type: object
      additionalProperties: true